feat(viewer): jump simulation for players + grenade rendering split
Simulate jumps in the 2D replay: derive a per-player "lift" (0..1) purely
from the Z coordinate (no parser change) using an asymmetric EMA ground
baseline that tells a jump apart from stairs/ramps. Z is sampled at 8/s and
not interpolated upstream, so interpolate it here to keep the lift smooth.
drawPlayer drops a ground shadow and lifts the body off it with a subtle
scale-up; labels follow the offset without scaling.

Fix grenade draw order: render ground decals (HE/molotov scorch) before the
live smoke/fire effects, so a lingering smoke is no longer covered by an HE's
dark patch that went off at the same spot.

Refactor: extract grenade rendering (effects, scorch, flight/preview arcs,
PATH_COLOR/KIND_ICON, effect radii) into a useGrenadeRendering composable.
Dependencies are injected (ctx getter, transform helpers, assets), call sites
are unchanged. ViewerMap.vue drops ~210 lines net.
